+++ Pinta.Core/Managers/WorkspaceManager.cs
@@ -204,8 +204,8 @@ namespace Pinta.Core
                 });
 
 				fileOpened = true;
-			} catch (UnauthorizedAccessException e) {
-				ShowOpenFileErrorDialog (parent, file, Catalog.GetString ("Permission denied"), e.ToString ());
+			} catch (UnauthorizedAccessException) {
+				ShowFilePermissionErrorDialog (parent, file);
 			} catch (FormatException e) {
 				ShowUnsupportedFormatDialog (parent, file, e.Message, e.ToString());
 			} catch (Exception e) {
@@ -370,6 +370,19 @@ namespace Pinta.Core
 
 			string message = string.Format (markup, primaryText, secondaryText);
 			PintaCore.Chrome.ShowUnsupportedFormatDialog(parent, message, details);
+		}
+
+		private void ShowFilePermissionErrorDialog (Window parent, string filename)
+		{
+			string markup = "<span weight=\"bold\" size=\"larger\">{0}</span>\n\n{1}";
+			string primary = Catalog.GetString ("Failed to open image");
+			// Translators: {0} is the name of a file that the user does not have permission to open.
+			string secondary = string.Format(Catalog.GetString ("You do not have access to '{0}'."), filename);
+			string message = string.Format (markup, primary, secondary);
+
+			var md = new MessageDialog (parent, DialogFlags.Modal, MessageType.Error, ButtonsType.Ok, message);
+			md.Run ();
+			md.Destroy ();
 		}
 
 		#region Public Events